Xwayback: add timeout to `-terminate`
This sets a timeout of 3 seconds after server reset.  This means that on
disconnection of the last running X client Xwayland will wait 3 seconds
for a new client.  This makes WMs' having a "replace" functionality
work, and also means anything in xinitrc that uses the X server before
an actual WM runs will not break startup.

Signed-off-by: Ferass El Hafidi <funderscore@postmarketos.org>