xwayback: Do not use SIGTERM for process handling
Relying on sending SIGTERM does not work reliably in case of SIGKILL or
crashes.

Instead, after starting wayback-compositor and Xwayland, just wait for
wayback-compositor to exit, and then exit ourselves. This allows all
three to cleanly exit by relying on each process to exit on its own when
the connection to another process is severed (waitid or socket).